refactor: publish settings
