+++ src/admin/pages/Post/PostFooter/PostFooterContent/Settings/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { LocalizedPost } from '../../../../../../types/index.js';
+
+export type Props = {
+  post: LocalizedPost;
+};
+
+export const Settings: React.FC<Props> = ({ post }) => {
+  return <>{post.title}</>;
+};
